vision-detection-classification/serverAI/requirements.txt
fastapi
paho-mqtt
uvicorn[standard]
pillow
moondream==0.0.6
onnxruntime
python-multipart
requests
